requirements.txt
pyrtlsdr
numpy
